Create a simple webserver
It prints out a unique machine ID when starting. The idea is that we can
use this ID to differentiate diff machine events. Though maybe this
isn't the correct place for that. Lets see.

I'm also not committing the vendor directory. Just to see how it goes.
I hate how the diffs become so much larger with it.