Try outputing everything in one file using tsc
The number of input files we will soon have will make it quite annoying
to have to list all of them in the manifest.json. I'm therefore trying
to use typescript to create just a single file. This kinda works, but
only with 'amd' or 'system' imports and when loading the extension, I
get error about a missing 'require' or 'system' function. Not sure how
to get this to work.